Fix: set consistent seeds (#59)
* fix: improve seeding for reproducible data

- adds a `safeSeed` function to handle strings that should be integers,
  e.g. IDs, which should be used in place of `faker.seed`!
- `src/factories/common.js` now exports a single version of `faker`,
  with the correct locale, this reduces possibility of hard to find
  errors where a differenet faker instance/locale has been seeded!
- persists land data once created
- standardises messages factory code, and ensures users' messages are
  different by adding the `personId` to the faker seed

* test: fix tests, simplify fixtures

The new `safeSeed` creates different data in the land factory, this
updates to the new content.
Some of the geometries (especially) in the land fixtures were verbose,
but offered little test value, this simplifies the assertions a bit.

* chore: add spelling exceptions

* fix: add tests, correct person update bug

Adds better tests to check the person update code works.
Person update implementation changed to NOT allow updating `emailValidated`.

* refactor: vanilla check host env var

Use the Node in-built `new URL(...)` funciton to check the `HOST` env
var value is valid config.
This removes the need for the previous implementation which used the
`convict-format-with-validator` module 🧹

* fix: add CRN to auth data

Reuse the generic `checkId` function to validate the CRN.
Ensure the CRN is returned in the authenticate data, NOT the personId!

* build: remove unused module dependency

Fully fixes all remaining issues reported by `npm audit` 🎉